Add PyTorch 1.12 and 1.13 to GH Actions (#478)
* Add PyTorch 1.12 and 1.13 to GH Actions

* Bump actions/setup-python from 4.4.0 to 4.5.0

Bumps [actions/setup-python](https://github.com/actions/setup-python) from 4.4.0 to 4.5.0.
- [Release notes](https://github.com/actions/setup-python/releases)
- [Commits](https://github.com/actions/setup-python/compare/v4.4.0...v4.5.0)

---
updated-dependencies:
- dependency-name: actions/setup-python
  dependency-type: direct:production
  update-type: version-update:semver-minor
...


* Format yaml configs

* Apply suggestions from @ialdencoots

* Temporarily disable test_onnx_export_single_image

+++ yolort/models/box_head.py
@@ -238,7 +238,7 @@ class SetCriterion(nn.Module):
 
         num_targets = targets.shape[0]
 
-        gain = torch.ones(7, device=device)  # normalized to gridspace gain
+        gain = torch.ones(7, device=device).long()  # normalized to gridspace gain
         # same as .repeat_interleave(num_targets)
         ai = torch.arange(num_anchors, device=device).float().view(num_anchors, 1).repeat(1, num_targets)
         # append anchor indices
